requirements.txt
# requirements.txt

# --- API & Web Server ---
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
httpx
redis

# --- LLM & AI ---
google-generativeai
openai  # <-- ADD THIS FOR DEEPSEEK

# --- Core Data Handling & Numerics ---
pandas
numpy

# --- Machine Learning ---
scikit-learn
xgboost
joblib

# --- Text Processing (Optional but good to have) ---
langchain-text-splitters